Taken all logic away from main
